# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: number;
  text: string;
  done: boolean;
  createdAt: string;
}

export interface CreateTaskBody {
  text: string;
}

export interface UpdateTaskBody {
  text?: string;
  done?: boolean;
}

export interface TaskStats {
  total: number;
  completed: number;
  pending: number;
}